Filter styles coming from the EditorProvider, not before it's passed to EditorProvider
