+++ packages/e2e-tests/plugins/iframed-enqueue-block-editor-settings.php
@@ -10,7 +10,10 @@
 add_action(
 	'block_editor_settings_all',
 	function( $settings ) {
-		$settings['styles'][] = array( 'css' => 'p { border: 1px solid red }' );
+		$settings['styles'][] = array(
+			'css' => 'p { border: 1px solid red }',
+			'__unstableType' => 'plugin',
+		);
 		return $settings;
 	}
 );
